Make package names singular
