+++ options-design/mongo/dbopt/dbopt.go
@@ -1,4 +1,4 @@
-package dbopts
+package dbopt
 
 type RunCommand interface {
 	runCommand()
