+++ options-design/mongo/mongoopt/mongoopt.go
@@ -1,4 +1,4 @@
-package mongoopts
+package mongoopt
 
 // Collation allows users to specify language-specific rules for string comparison, such as
 // rules for lettercase and accent marks.
